[TIR] Use IndexMap to transform NDArray (#12949)
I've hit a weird use case where I want to manually transform `runtime::NDArray` (attached to `AllocateConst` node) according to the index map used in `transform_layout`. This is needed to support `AllocateConst` node in Metaschedule `RewriteLayout` postproc.

I can define it as a free function in the file where it is actually used. Having it available as part of the `IndexMap` interface makes it convenient to expose this to python and unit-test it. Let me know if this is a reasonable API addition.