[Arith] Optional rewriting and simplification into AND of ORs (#12972)
Previously, `RewriteSimplifier` inspected the top branch of each `And`
and `Or` node, but didn't perform simplifications that would require
inspection across branches nested structures.  This commit introduces
`SimplifyAsAndOfOrs`, which converts to an internal representation as
conjunctive normal form, performs simplifications while in that form,
then converts back to a `PrimExpr`.

This utility is designed for data-propagation simplifications as part
of https://github.com/apache/tvm/issues/12261.  To avoid increasing
the CPU load unnecesarily, this utility is only used on an opt-in
basis, either using `RewriteSimplifer::SetEnabledFeatures` when called
directly, or using `PassContext` when used as part of
`tir::transform::Simplify`.

* [Arith][UnitTest] Unittests displaying desired behavior

* Corrected example in analyzer.h

* Added underscore for private members

* Rename Cleanup() to RemoveTrueFalse()

* Added comments for conversion to internal representation

* Updated comment in test case

* Preferentially order clauses according to first occurrence

* Added more unit tests