[Hexagon] [runtime] Per-thread hardware resource management (#13181)
Add support to lock and unlock resources on the individual threads. Adds APIs to get a thread handle for a given resource type, and get a resource type for a given thread handle. This will aid in asynchronous support that is sending work to dedicated threads for the hardware resources.

There are two steps to fully acquire a resource: reserve and lock. HTP/HVX manager objects are created on the global thread, which will reserve the resources. The lock/unlock will occur on the thread that owns a particular instance. The manager objects will be deconstructed when the thread manager is deconstructed. This will release the resources.