[microTVM] Update Zephyr version and Zephyr SDK version (#13818)
This commit updates the microTVM code to use Zephyr 3.2 and SDK 0.15.2.
                                                                       
As a result of this change, there are a few other changes that are     
included:           

- A launch script was added for Zephyr and Arduino template project to 
run the Project API server from a different Python ENV than TVM Python 
ENV;

- For Zephyr, the launch script uses global python3.8 which is where   
west is registered. However, for Arduino it uses a separate virtual ENV
with python3 version that exists in the host;
                                                                       
- tests/micro/project_api/test_arduino_microtvm_api_server.py was      
removed since these tests were using Arduino microTVM API server by    
importing it from TVM. We no longer support Arduino/Zephyr dependencies
in TVM testing python ENV;
                                                                       
- Disables a demo and test due to a CMSIS-NN bug:                      
[Bug] CMSIS-NN BYOC fails with Zephyr 3.2 #13856;                      
                
There will be a follow up work to move Zephyr to a completely separate 
virtual ENV as it was done in this commit for Arduino in the launch
script.