[Unity][Transform] Memory plan across the IRModule (#14220)
Previously the static memory planning pass only works at single function
level - each function inside the an IRModule will be independently
planned. This is not perfect for the VM to reuse allocated memory across
different functions.

Therefore, this PR turns the static memory planning pass into a module
pass. Now the plan is done across the IRModule, so that memory alloation
in different functions can share the same storage token when planning.
With this PR, it is hopeful that the VM will find more opportunities of
memory reuse.