[Schedule] Add an optional argument `disable_checks` for `Schedule` (#14281)
# Motivation
Currently, some of the schedule checks are too strict, which makes it hard to schedule some workloads such as FlashAttention whose reduction is two-stage and does not strictly follows our standard.

This PR adds an optional argument `disable_checks` which mutes some checks. The argument defaults to `False` and we can enable it whenever we want to disable some `soft` checks (by `soft` we mean if we violate such checks, the schedule is not necessarily invalid, and if we violate `hard` checks the schedule step is invalid).

In the future, we should collect the `soft` and `hard` checks for all schedule primitives. This PR serves for FlashAttention and only cares `bind` and some reduction primitives for now.