[Codegen][LLVM] Remove cast to i8* in builtin::address_of (#14563)
* [Codegen][LLVM] Remove cast to i8* in builtin::address_of

This cast was initially added when `CreateBufferPtr` did not include
the cast to the appropriate address space, and is no longer
necessary.  Removing the cast will not harm LLVM configurations with
opaque pointers ([default in LLVM
15+](https://github.com/llvm/llvm-project/blob/main/llvm/docs/OpaquePointers.rst)),
and avoids type mismatches when using LLVM configurations that use
typed pointers.

* Added handling of i8* argument to non-overloaded prefetch

For compatibility with earlier versions of LLVM.  Newer versions have
an overloaded prefetch intrinsic, and the pointer cast is unnecessary.