[Unity] Fix FuseTIR when the same buffer is read multiple times with different access pattern (#14603)
When the same buffer is read multiple times with different access patterns in a single expression, the check below fails

But this case should be allowed, for example in the following subgraph inp_0 is used twice in different read regions. See the test case for details. This subgraph arises if we run ConvertLayout on SD UNet from web-stable diffusion.