[Unity][Training] Trainer and SetupTrainer (#14706)
[Training] Loss function, Optimizer, SetupTrainer, and Trainer
This PR adds two utilities to simplify the training process: Trainer and SetupTrainer.

SetupTrainer is basically a pass, transforming a backbone module to a complete, legalized trainer module. It is a combination of Gradient pass, loss function, optimizer, AppendLoss pass, LegalizeOps, etc. You can store the trainer module into a file and load it to somewhere else.

Trainer is a runtime utility that accepts the IRModule (that is the result of SetupTrainer) and the relax VM (that contains the built result of the IRModule), and helps run the VM. It maintains the parameters, the model states and the optimizer states internally.



Co-authored-by: Chaosfan <siriusneo@sjtu.edu.cn>