[TVMScript] Use default fallback for un-registered type (#14874)
Previously, even if an object is supported by TVMScript, printing that
object could fail if the object contains references to un-registered
objects.  In particular, the attributes in the `PrimFunc` and
`AttrStmt` may contain any `ObjectRef`, and are frequently used for
external bookkeeping.  This commit adds a default printer for
unregistered types in TVMScript.