[TIR] Expand unit tests for ConvertSSA (#14892)
* [TIR] Expand unit tests for ConvertSSA

Prior to this PR, there was a single test which invoked ConvertSSA and
checked that no error was thrown.  This PR adds additional test cases
for nested variable definition, and for variables de-duplicated across
separate function calls in an `IRModule`.

Of the additional tests, the behavior tested by
`TestDedupAutoBroadcastBuffer` and `TestReusedBufferParameter` fails
on main, and is resolved by this PR.

* Update tests to avoid relying on TVMScript output of non-SSA

The "before" cases must have SSA violations, which is not valid in
TIR.  As such, future versions of the TVMScript parser may remove the
SSA violations in the process of parsing.  Updated unit tests
introduce SSA violations through the Python API, to avoid this
potential breakage.