[Unity][Relax] Make RewriteDataflowReshape only rewrite volume-preserving ops (#15112)
* [Unity] Make RewriteDataflowReshape only rewrite volume-preserving ops

The reshape operator expects that the number of elements in the source
is the same as the number of elements in the result. There are operators
that could have a reshape pattern that don't meet this requirement (e.g.
strided_slice), and they should not be converted to reshape.

* Move shape verification to IsCallingTIRReshape

* Replace ICHECK_EQ(used_arg_indices.size(), 1) with return

Since the check for has-reshape-pattern is done after this check, so
don't abort if the check fails, just return.