[TIR] Call TVMBackendFreeWorkspace inside LetStmt (#15253)
* [TIR] Call TVMBackendFreeWorkspace inside LetStmt

Prior to this commit, the call to `TVMBackendFreeWorkspace` occurred
outside the `LetStmt` that defined the workspace pointer.  While works
with current codegen, as the code produced for `LetStmt` does not
check for out-of-scope access, this access of an out-of-scope should
be avoided.

This commit updates `LowerTVMBuiltin` to produce the call to
`TVMBackendFreeWorkspace` at the end of the `LetStmt`'s body, rather
than just after the `LetStmt`.

* [TIR] Output AttrStmt "storage_alignment" inside the var binding

Prior to this commit, the `AttrStmt` providing the storage alignment
was placed outside the `LetStmt` that defines the variable. As a
result, the alignment assumption is never actually used, as
`CodeGenLLVM::VisitStmt_(const AttrStmtNode*)` only creates an
alignment assumption for in-scope variables.

This commit moves the storage alignment `AttrStmt` to be inside the
`LetStmt`, rather than outside.