[Unity][NN] Allow nn.Module to generate subroutines (#15453)
* [Relax][IRBuilder] Allow subroutine definition in BlockBuilder

Prior to this commit, a user needed to first define any subroutines
used within a module, then define the functions that use those
subroutines.  This prevents some use cases, such as
`tvm.relax.testing.nn.Module` implementations that generate a
subroutine when required.

This commit updates the `tvm.relax.block_builder` to maintain a stack
of functions currently being defined, such that subroutines can be
defined as needed, resuming collection into the main function
afterwards.

* [Unity][NN] Allow nn.Module to generate subroutines

Prior to this commit, `nn.Module` generated the relax expression
within the body of `relax.BlockBuilder.current()`.  For large models,
this can result in extremely large function bodies, making it
difficult to identify regions within the module.

This commit adds an option `nn.Module.define_subroutine`, which
defaults to `False` (current behavior).  If set to `True`; either
within a subclass, within an instance, or globally in `nn.Module`;
function calls into the module will produce a subroutine representing
the module's execution, and a call into that subroutine.  For example,
calling a `class Linear(nn.Module)` would produce a `def linear(arg,
weights)` function definition and a `module.linear(arg, weights)`
function call.

To ensure correct shape propagation, a subroutine is generated for
each unique set of argument shapes passed to the `nn.Module` subclass.

* Backwards compatibility workarounds

If a `def forward` expected a python-style tuple object, it may now
receive a `relax.Tuple`, or a `relax.Var` annotated with
`relax.TupleStructInfo`.  This commit adds backwards compatibility
type checks to allow common python operations to keep the same
result in these cases.

* Update accessor of _blocks

* [Relax][NN] Implemented subroutines for relax.frontend.nn

* [Bugfix][Relax] In-scope late-bound parameters

Prior to this commit, the normalization of a function body was
performed with only the early-binding parameters in-scope, but not the
late-binding parameters.  Therefore, if a parameter has a dynamic
shape parameter that determines a dynamic output shape, the output
shape would only be inferred correctly if the parameter is provided as
part of `block_builder.function(name, params=[...])`.  If the
parameter is instead provided as part of
`block_builder.emit_func_output(output, params=[...])`, the output
would be erroneously annotated with `R.Tensor(ndims=ndims)` instead of
`R.Tensor(shape=dynamic_shape)`.

This commit updates `BlockBuilder.emit_func_output` to ensure the
parameters are in-scope, regardless of when they are provided.

* Updated implementation in relax.testing.nn

* Lint fixes