[Unity] Propagate dynamic begin/end in strided_slice when legalizing (#15515)
* [Unity] Propagate dynamic begin/end in strided_slice when legalizing

A follow-up to PR #15450, which allowed dynamic begin/end expressions
in `relax.op.strided_slice`.  If the shape of the output is static,
such as a slice from `index:index+slice_size`, the relax operator will
delegate to `topi.strided_slice`, but that may then delegate to
`topi.dynamic_strided_slice` due to the dynamic begin/end.  In this
case, the topi implementation inserts an additional output variable
`"dim"`, without providing a definition.

This behavior works in topi, as the output shape variable may have a
definition provided from the calling scope, when the output tensor is
passed in.  However, even in topi, this behavior is fragile as it
assumes that the Tensor output is the same as the resulting PrimFunc
output.

This commit updates the topi implementation of `strided_slice` to
propagate the shape information from the inputs.

* Updated unit test