[Relay] Disable exception for ADT in mixed precision pass (#15533)
If topology contains while loop and we want to transform it to mixed
precision then we get an exception that "ADT are not supported for
mixed precision pass". It happens, because while loop implemented as a
lambda which is assigned to a VarNode.

In this commit I changed the behavior of ToMixedPrecision pass and
instead of generating exception, it just do nothing.

Correspondent regression test is added.