[Arith] Fix handling of overlapping predicates (#15555)
When the lower bound predicate is present, `TryFuseIters` will rewrite
an expression to `IterMarkWithOffset`, where the iter mark (structured
form) has minimum value of zero. During rewriting, when the subexpression
matches a previous rewritten result (`IterMarkWithOffset`), previously
it doesn't correctly add the offset to the result `IterMarkWithOffset`,
and causes inconsistency between the flattened form and the structured
(normal) form.