[Unity][Frontend][NN] Make effects optional in nn module. (#15650)
This PR adds the `debug` argument to `export_tvm`. When `debug` is `False`, effects are not included in the output graph. This can make deploying models less cumbersome since its not often theyll use effects. I also added automatic annotation of the `num_inputs` function attribute since it is useful for a few passes.