[Unity][Analysis] Check for usage of DataflowVar in all_vars() (#15698)
Prior to this commit, the `VarVisitor` used in the implementation of
`all_vars` and `free_vars` only collected variable usage sites where
the variable was a `const VarNode*`, and ignored usage sites of a
`const DataflowVarNode*`.  When analyzing an entire function, these
variables were found in the `const VisitVarBinding*`, and could pass
the existing tests.  However, when analyzing a single expression,
these variables would be erroneously excluded.

This commit adds a `VisitExpr_(const DataflowVarNode*)` implementation
in `VarVisitor`, to collect variable usage regardless of the type of
variable.